% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cap_outliers.R
\name{cap_outliers}
\alias{cap_outliers}
\title{cap_outliers function}
\usage{
cap_outliers(vector)
}
\arguments{
\item{vector}{A vector containing continuous values.}
}
\description{
cap_outliers function
}
