% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db_connection.R
\name{db_connection}
\alias{db_connection}
\title{db_connection function}
\usage{
db_connection(db_name, lib_path)
}
\arguments{
\item{db_name}{A character string. The name of the database.}

\item{lib_path}{A character string. The path to the utilities-folder.}
}
\description{
db_connection function
}
