% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/launch_app.R
\name{launch_app}
\alias{launch_app}
\title{Launch the labVisualizeR}
\usage{
launch_app(
  port = 3838,
  db_name,
  utils_path = system.file("application/_utilities", package = "labVisualizeR"),
  lib_path
)
}
\arguments{
\item{port}{The port, baseApp is running on (default: 3838)}

\item{db_name}{A character string. The name of the database.}

\item{utils_path}{A character string. The path to the utilities-folder.}

\item{lib_path}{A character string. The path to the utilities-folder.}
}
\value{
labVisualizeR application
}
\description{
Launch the labVisualizeR
}
