% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_sql.R
\name{load_sql}
\alias{load_sql}
\title{load_sql function}
\usage{
load_sql(db_name, utils_path)
}
\arguments{
\item{db_name}{A character string. The name of the database.}

\item{utils_path}{A character string. The path to the utilities-folder.}
}
\description{
load_sql function
}
