% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/moduleFrequency.R
\name{module_frequency_ui}
\alias{module_frequency_ui}
\title{module_frequency_ui}
\usage{
module_frequency_ui(id)
}
\arguments{
\item{id}{A character. The identifier of the shiny object}
}
\description{
module_frequency_ui
}
