% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/moduleMetadata.R
\name{module_metadata_ui}
\alias{module_metadata_ui}
\title{module_metadata_ui}
\usage{
module_metadata_ui(id)
}
\arguments{
\item{id}{A character. The identifier of the shiny object}
}
\description{
module_metadata_ui
}
