% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/moduleRawdata.R
\name{module_rawdata_ui}
\alias{module_rawdata_ui}
\title{module_rawdata_ui}
\usage{
module_rawdata_ui(id)
}
\arguments{
\item{id}{A character. The identifier of the shiny object}
}
\description{
module_rawdata_ui
}
