% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/moduleVisualize.R
\name{module_visualize_ui}
\alias{module_visualize_ui}
\title{module_visualize_ui}
\usage{
module_visualize_ui(id)
}
\arguments{
\item{id}{A character. The identifier of the shiny object}
}
\description{
module_visualize_ui
}
