% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reset.R
\name{reset}
\alias{reset}
\title{reset}
\usage{
reset(session, db_got_num, db_got_cat, db_data)
}
\arguments{
\item{session}{Shiny session object}

\item{db_got_num}{A logical, indicates if VALUE_NUM is present and the 
range slider needs to be populated.}

\item{db_got_cat}{A logical, indicates if only discrete data is present.}

\item{db_data}{The dataset}
}
\description{
reset
}
