% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reset_min_max.R
\name{reset_min_max}
\alias{reset_min_max}
\title{reset_min_max}
\usage{
reset_min_max(db_got_num, db_data_subset)
}
\arguments{
\item{db_got_num}{A logical, indicates if VALUE_NUM is present and the 
range slider needs to be populated.}

\item{db_data_subset}{The subset dataset}
}
\description{
reset_min_max
}
