% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary_function.R
\name{summary_function}
\alias{summary_function}
\title{summary_function function}
\usage{
summary_function(vector)
}
\arguments{
\item{vector}{A vector containing continuous values.}
}
\description{
summary_function function
}
