% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update_slider.R
\name{update_sliders}
\alias{update_sliders}
\title{update_sliders}
\usage{
update_sliders(
  session,
  db_got_num,
  range_min = NULL,
  range_max = NULL,
  min_age,
  max_age
)
}
\arguments{
\item{session}{Shiny session object}

\item{db_got_num}{A logical, indicates if VALUE_NUM is present and the 
range slider needs to be populated.}

\item{range_min}{The minimum value of VALUE NUM (default: NULL)}

\item{range_max}{The maximum value of VALUE NUM (default: NULL)}

\item{min_age}{The minimum age value of the underlying cohort.}

\item{max_age}{The maximum age value of the underlying cohort.}
}
\description{
update_sliders
}
